refactor: renomeado aos arquivos Log para Record e ajustado as referência
