+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { UploadModule } from './upload/upload.module';
-import { LogsModule } from './logs/logs.module'; // Certifique-se de importar o LogsModule
+import { RecordModule } from './record/record.module'; // Certifique-se de importar o RecordModule
 
 @Module({
   imports: [
@@ -13,7 +13,7 @@ import { LogsModule } from './logs/logs.module'; // Certifique-se de importar o 
       entities: [__dirname + '/**/*.entity{.ts,.js}'], // localização das entidades
       synchronize: true, // deve ser usado apenas em desenvolvimento
     }),
-    LogsModule, // Adicione o LogsModule
+    RecordModule, // Adicione o RecordModule
     UploadModule,
   ],
   controllers: [AppController],
