+++ src/record/record.controller.ts
@@ -1,0 +1,12 @@
+import { Controller, Get } from '@nestjs/common';
+import { RecordService } from './record.service';
+
+@Controller('record')
+export class RecordController {
+  constructor(private readonly logsService: RecordService) {}
+
+  @Get()
+  findAll() {
+    // Lógica para retornar record
+  }
+}
