+++ src/record/record.service.ts
@@ -1,0 +1,6 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class RecordService {
+  // Lógica do serviço
+}
