+++ src/upload/upload.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 import { UploadController } from './upload.controller';
 import { UploadService } from './upload.service';
 // import { MockUploadService } from './upload.mock.service';
-import { LogsModule } from '../logs/logs.module';
+import { RecordModule } from '../record/record.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([]), LogsModule],
+  imports: [TypeOrmModule.forFeature([]), RecordModule],
   controllers: [UploadController],
   providers: [UploadService],
 })
